% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_timer}
\alias{load_timer}
\title{Load a Timer database.}
\usage{
load_timer(path)
}
\arguments{
\item{path}{a path to the database file.}
}
\value{
a \code{\link{timer}} class object.
}
\description{
Loads a local or remote Timer immune infiltration database
(http://timer.cistrome.org/) and stores it as a timer class object.
}
