% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oop.R
\name{timer}
\alias{timer}
\title{Create a timer class object.}
\usage{
timer(x, ...)
}
\arguments{
\item{x}{an object of the tibble class.}

\item{...}{extra arguments, currently none specified.}
}
\value{
an object of the class timer.
}
\description{
Creates a timer object storing a database with the Timer 2.0
immune infiltration estimates.
}
\details{
Technically, inherits most of the methods from
\code{\link[tibble]{tibble}}. Has an attribute 'cell_types' storing the
available cell types.
}
